[NCL-4958] migrated bpm-dependant rest-model entities to bpm module, reversed bpm->rest-model dependency, fixed links and tests
