+++ bpm/pom.xml
@@ -41,10 +41,6 @@
     </dependency>
     <dependency>
       <groupId>org.jboss.pnc</groupId>
-      <artifactId>rest-model</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.pnc</groupId>
       <artifactId>common</artifactId>
     </dependency>
     <dependency>
