+++ bpm/src/main/java/org/jboss/pnc/bpm/model/RepositoryCreationResultRest.java
@@ -15,18 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.restmodel.bpm;
-
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.jboss.pnc.enums.BPMTaskStatus;
+package org.jboss.pnc.bpm.model;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.ToString;
 
 import javax.xml.bind.annotation.XmlRootElement;
@@ -34,38 +31,37 @@ import javax.xml.bind.annotation.XmlRootElement;
 /**
  * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
  */
-@JsonDeserialize(builder = ProcessProgressUpdate.ProcessProgressUpdateBuilder.class)
+@JsonDeserialize(builder = RepositoryCreationResultRest.RepositoryCreationResultRestBuilder.class)
 @AllArgsConstructor
 @Builder
-@XmlRootElement
+@XmlRootElement(name = "RepositoryCreationResultRest")
 @ToString
 @NoArgsConstructor
 @Setter(onMethod=@__({@Deprecated}))
-public class ProcessProgressUpdate extends BpmNotificationRest {
-
-    /**
-     * Name of the service managed by the BPM eg. Repour
-     */
-    @Getter
-    private String taskName;
-
-    @Getter
-    private BPMTaskStatus bpmTaskStatus;
-
-    /**
-     * Url to subscribe to detailed notification.
-     * Notifications can be a string stream of live log or an object with detailed statuses.
-     */
-    @Getter
-    private String detailedNotificationsEndpointUrl;
+public class RepositoryCreationResultRest extends BpmNotificationRest {
 
     @Override
     public String getEventType() {
-        return "PROCESS_PROGRESS_UPDATE";
+        return eventType.name();
     }
+
+    @Getter
+    private Integer repositoryId;
+
+    @Getter
+    private Integer buildConfigurationId;
+
+    @Getter
+    private EventType eventType;
+
+    @Getter
+    private String errorMessage;
 
     @JsonPOJOBuilder(withPrefix = "")
-    public static final class ProcessProgressUpdateBuilder {
+    public static final class RepositoryCreationResultRestBuilder {
     }
 
+    public enum EventType {
+        RC_CREATION_SUCCESS, RC_CREATION_ERROR;
+    }
 }
