+++ bpm/src/main/java/org/jboss/pnc/bpm/model/causeway/ArtifactImportError.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.restmodel.causeway;
+package org.jboss.pnc.bpm.model.causeway;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
