+++ bpm/src/main/java/org/jboss/pnc/bpm/model/causeway/BuildImportResultRest.java
@@ -15,21 +15,50 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.restmodel.causeway;
+package org.jboss.pnc.bpm.model.causeway;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 /**
  * Author: Michal Szynkiewicz, michal.l.szynkiewicz@gmail.com
  * Date: 8/25/16
- * Time: 2:22 PM
+ * Time: 2:48 PM
+ *
  */
-public enum BuildImportStatus {
+@Deprecated
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@NoArgsConstructor
+@AllArgsConstructor
+public class BuildImportResultRest {
     /**
-     * setting up the import process failed
+     * id of pnc build record
      */
-    ERROR,
+    private Integer buildRecordId;
+
     /**
-     * some of the artifacts failed to import
+     * build id assigned by brew
      */
-    FAILED,
-    SUCCESSFUL
+    private Integer brewBuildId;
+    /**
+     * link to brew
+     */
+    private String brewBuildUrl;
+
+    private BuildImportStatus status;
+    /**
+     * global errors
+     */
+    private String errorMessage;
+
+    /**
+     * list of errors for artifact imports
+     */
+    private List<ArtifactImportError> errors;
+
 }
