+++ bpm/src/main/java/org/jboss/pnc/bpm/task/BpmBuildTask.java
@@ -20,12 +20,11 @@ package org.jboss.pnc.bpm.task;
 import lombok.ToString;
 import org.jboss.pnc.bpm.BpmManager;
 import org.jboss.pnc.bpm.BpmTask;
+import org.jboss.pnc.bpm.model.BuildExecutionConfigurationRest;
+import org.jboss.pnc.bpm.model.ComponentBuildParameters;
 import org.jboss.pnc.common.util.TimeUtils;
-import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationAudited;
 import org.jboss.pnc.model.utils.ContentIdentityManager;
-import org.jboss.pnc.rest.restmodel.BuildExecutionConfigurationRest;
-import org.jboss.pnc.rest.restmodel.bpm.ComponentBuildParameters;
 import org.jboss.pnc.spi.coordinator.BuildTask;
 import org.jboss.pnc.spi.exception.CoreException;
 import org.jboss.pnc.spi.executor.BuildExecutionConfiguration;
