+++ bpm/src/test/java/org/jboss/pnc/bpm/test/BpmSchedulerSmokeTest.java
@@ -21,8 +21,8 @@ package org.jboss.pnc.bpm.test;
 import lombok.Getter;
 import org.jboss.pnc.bpm.BpmManager;
 import org.jboss.pnc.bpm.BpmTask;
+import org.jboss.pnc.bpm.model.BpmStringMapNotificationRest;
 import org.jboss.pnc.common.json.moduleconfig.BpmModuleConfig;
-import org.jboss.pnc.rest.restmodel.bpm.BpmStringMapNotificationRest;
 import org.jboss.pnc.spi.exception.CoreException;
 import org.junit.Before;
 import org.junit.Test;
