--- rest-model/src/main/java/org/jboss/pnc/rest/restmodel/bpm/ComponentBuildParameters.java
@@ -1,65 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2019 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.rest.restmodel.bpm;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.jboss.pnc.rest.restmodel.BuildExecutionConfigurationRest;
-
-import javax.validation.constraints.NotNull;
-import java.io.Serializable;
-
-/**
- * Component Build process input parameters
- */
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class ComponentBuildParameters implements Serializable {
-
-    @NotNull
-    @Getter
-    private String pncBaseUrl;
-
-    @NotNull
-    @Getter
-    private String aproxBaseUrl;
-
-    @NotNull
-    @Getter
-    private String repourBaseUrl;
-
-    @NotNull
-    @Getter
-    private String daBaseUrl;
-
-    @NotNull
-    @Getter
-    private Boolean communityBuild;
-
-    @NotNull
-    @Getter
-    private Boolean versionAdjust;
-
-    @NotNull
-    @Getter
-    BuildExecutionConfigurationRest buildExecutionConfiguration;
-
-}
