+++ rest/src/test/java/org/jboss/pnc/rest/provider/MilestoneReleaseTest.java
@@ -23,10 +23,10 @@ import org.jboss.pnc.model.BuildRecord;
 import org.jboss.pnc.model.ProductMilestone;
 import org.jboss.pnc.model.ProductMilestoneRelease;
 import org.jboss.pnc.rest.restmodel.ProductMilestoneRest;
-import org.jboss.pnc.rest.restmodel.causeway.ArtifactImportError;
-import org.jboss.pnc.rest.restmodel.causeway.BuildImportResultRest;
-import org.jboss.pnc.rest.restmodel.causeway.BuildImportStatus;
-import org.jboss.pnc.rest.restmodel.causeway.MilestoneReleaseResultRest;
+import org.jboss.pnc.bpm.model.causeway.ArtifactImportError;
+import org.jboss.pnc.bpm.model.causeway.BuildImportResultRest;
+import org.jboss.pnc.bpm.model.causeway.BuildImportStatus;
+import org.jboss.pnc.bpm.model.causeway.MilestoneReleaseResultRest;
 import org.jboss.pnc.enums.ReleaseStatus;
 import org.jboss.pnc.rest.utils.mock.BpmMock.Push;
 import org.jboss.pnc.rest.utils.mock.BpmMock.PushList;
