+++ rest/src/test/java/org/jboss/pnc/rest/utils/mock/BpmMock.java
@@ -26,8 +26,8 @@ import org.jboss.pnc.common.Configuration;
 import org.jboss.pnc.common.json.ConfigurationParseException;
 import org.jboss.pnc.common.json.moduleconfig.BpmModuleConfig;
 import org.jboss.pnc.common.json.moduleprovider.ConfigProvider;
-import org.jboss.pnc.rest.restmodel.bpm.MilestoneReleaseParameters;
-import org.jboss.pnc.rest.restmodel.causeway.MilestoneReleaseRest;
+import org.jboss.pnc.bpm.model.MilestoneReleaseParameters;
+import org.jboss.pnc.bpm.model.causeway.MilestoneReleaseRest;
 import org.jboss.pnc.spi.exception.CoreException;
 import org.jboss.pnc.test.util.JsonUtils;
 import org.kie.api.definition.process.Process;
